"Revised the code for the issue."
I revised the whole code for possible typos or errors when translating the code from Python, but the issue with the ground state energy is still there. I even tried a simple case on both Python and Julia, but the rng work different on these languages so after a couple iterations they show much different results.

The issue with the Julia version is that the learning process is very "noisy", that is the energy decreases very slowly and with a lot of noise. In addition, increasing the number of iterations further makes the energy diverge, which is very odd, since in Python this does not happen.